exp_1/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pika==1.3.2
redis==5.0.1
mysql-connector-python==8.3.0